Add google command
A command that searches google. 
Usage: <prefix>google <query> [results: <number of results to display>] [--send]

The --send fag will send the result to a channel for everyone to see. If this flag isn't specified, it will instead be a client-side message.
The results: n parameter allows you to specify how many results you want to see. Note that if you have the --send flag, this will be capped to one. Defaults to three.