Add google command to ExternalCommands (#57)
* Add google command

A command that searches google. 
Usage: <prefix>google <query> [results: <number of results to display>] [--send]

The --send fag will send the result to a channel for everyone to see. If this flag isn't specified, it will instead be a client-side message.
The results: n parameter allows you to specify how many results you want to see. Note that if you have the --send flag, this will be capped to one. Defaults to three.

* Add dependency for google

See 7e35d8b4699619db627f179da8d98d96458f7b6f

* Update google.js result structure

The `Search results for ${args}` is now moved to message.content, and the embed title is now the title belonging to the result. This is better because there's way less repetition and wasted space, and it looks cleaner.